# src/api/openapi.yml
openapi: 3.0.0
info:
  version: 0.1.0
  title: AgentENV API
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    # AuthProviderBearerAuth / AuthProviderTeamAuth: B before T in the name
    # so Bearer is validated before Team.
    AuthProviderBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    AuthProviderTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
    # AdminApiKeyAuth / AdminTeamAuth: alphabetical names ensure token
    # validation runs before team context population.
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    AdminTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
  parameters:
...
    sandboxID:
      name: sandboxID
      in: path
      required: true
      schema:
        type: string
...
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "409":
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "500":
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    EnvdVersion:
      type: string
      description: Version of the envd running in the sandbox
...
    Sandbox:
      required:
        - templateID
        - sandboxID
        - clientID
        - envdVersion
      properties:
        templateID:
          type: string
          description: Identifier of the template from which is the sandbox created
        sandboxID:
          type: string
          description: Identifier of the sandbox
        alias:
          type: string
          description: Alias of the template
        clientID:
          type: string
          deprecated: true
          description: Identifier of the client
        envdVersion:
          $ref: "#/components/schemas/EnvdVersion"
        envdAccessToken:
          type: string
          description: Access token used for envd communication
        trafficAccessToken:
          type: string
          nullable: true
          description: Token required for accessing sandbox via proxy.
        domain:
          type: string
          nullable: true
          description: Base domain where the sandbox traffic is accessible
...
    ConnectSandbox:
      type: object
      required:
        - timeout
      properties:
        timeout:
          description: Timeout in seconds from the current time after which the sandbox should expire
          type: integer
          format: int32
          minimum: 0
    ConnectSandboxV2:
      type: object
      properties:
        timeout:
          description: Timeout in seconds from the current time after which the sandbox should expire
          type: integer
          format: int32
          minimum: 1
          default: 300
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error
...
paths:
...
  /sandboxes/{sandboxID}/connect:
    post:
      summary: Connect sandbox
      description: Returns sandbox details. If the sandbox is paused, it will be resumed. TTL is only extended. Use POST /v2/sandboxes/{sandboxID}/connect instead.
      deprecated: true
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/sandboxID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectSandbox"
      responses:
        "200":
          description: The sandbox was already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "201":
          description: The sandbox was resumed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
  /v2/sandboxes/{sandboxID}/connect:
    post:
      summary: Connect sandbox (v2)
      description: >-
        Returns sandbox details. If the sandbox is paused, it will be resumed.
        TTL is only extended. The request body is optional; an omitted timeout
        defaults to 300 seconds.
      tags: [sandboxes]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/sandboxID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectSandboxV2"
      responses:
        "200":
          description: The sandbox was already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "201":
          description: The sandbox was resumed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
...